requirements.txt

python-telegram-bot>=20.7
openai>=1.6.1,<2.0.0
langchain>=0.1.0,<1.0.0
langchain-core>=0.1.0,<1.0.0
langchain-community>=0.0.10,<1.0.0
langchain-text-splitters>=0.0.1,<1.0.0
langchain-openai>=0.0.2
zhipuai>=1.0.0
dashscope>=1.14.0
minimax>=0.0.2
sqlalchemy>=2.0.0
apscheduler>=3.10.0
python-dotenv>=1.0.0
aiohttp>=3.9.0
pytz>=2023.3
tiktoken>=0.5.0
psutil>=7.2.2
pynput>=1.8.1
pyobjc>=12.1
